client: move kb-bankpin back into bank plugin
